python_client/requirements.txt
pyserde==0.14.0
websockets==8.1
